refactor(robot): renomme roblaude_arm en roblaude_pickplace
