--- robot/roblaude_arm/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/roblaude_arm
-[install]
-install_scripts=$base/lib/roblaude_arm
