+++ robot/roblaude_nav/package.xml
@@ -24,7 +24,7 @@
   <exec_depend>nav2_msgs</exec_depend>
   <exec_depend>action_msgs</exec_depend>
   <exec_depend>std_msgs</exec_depend>
-  <exec_depend>roblaude_arm</exec_depend>
+  <exec_depend>roblaude_pickplace</exec_depend>
   <exec_depend>arm_msgs</exec_depend>
   <exec_depend>foxglove_bridge</exec_depend>
   <exec_depend>web_video_server</exec_depend>
