+++ robot/roblaude_pickplace/config/detection_params.yaml
