+++ robot/roblaude_pickplace/launch/pickplace.launch.py
@@ -5,7 +5,7 @@ par roblaude_nav/camera.launch.py — on evite ainsi une dependance croisee entr
 les deux packages. L'autostart du container lance les deux (decision C : camera
 + detecteur toujours ON).
 
-    ros2 launch roblaude_arm pickplace.launch.py
+    ros2 launch roblaude_pickplace pickplace.launch.py
 """
 from launch import LaunchDescription
 from launch.substitutions import PathJoinSubstitution
@@ -15,11 +15,11 @@ from launch_ros.substitutions import FindPackageShare
 
 def generate_launch_description():
     params = PathJoinSubstitution([
-        FindPackageShare('roblaude_arm'), 'config', 'detection_params.yaml'])
+        FindPackageShare('roblaude_pickplace'), 'config', 'detection_params.yaml'])
 
     return LaunchDescription([
         Node(
-            package='roblaude_arm',
+            package='roblaude_pickplace',
             executable='object_detector',
             name='object_detector',
             parameters=[params],
