+++ robot/roblaude_pickplace/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>roblaude_arm</name>
+  <name>roblaude_pickplace</name>
   <version>0.1.0</version>
   <description>
     Pick and place RobLaude (UC-02) pour le bras du ROSMASTER M3 PRO.
