+++ robot/roblaude_pickplace/roblaude_pickplace/detection.py
