+++ robot/roblaude_pickplace/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/roblaude_pickplace
+[install]
+install_scripts=$base/lib/roblaude_pickplace
